+++ CHANGELOG.md
@@ -5,6 +5,9 @@
     - Removed right padding/margin for last items in `hstack` and `columns`
     - Changed styles for `button`, `input` and `select` to be consistent with each others
     - Added disabled state for `menu-item` view
+    - Added a message and disable menu items when JSON can't be copied through `struct` actions popup
+    - Improved experience when JSON can't be copied through `struct` view actions, items are disabled and annotated with a reason message now
+    - Added size of compacted JSON in `struct` actions menu
 
 ## 1.0.0-beta.11 (10-02-2019)